l3gd20: split into separate header, implementation, and main
